% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTmerge.R
\name{DTmerge}
\alias{DTmerge}
\title{Merge two data.tables by like columns}
\usage{
DTmerge(x, y, all = F, all.x = F, all.y = F)
}
\arguments{
\item{x}{a data.table}

\item{y}{a data.table}
}
\value{
a merged data.table
}
\description{
Merge two data.tables by like columns
}
