% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTnull.R
\name{DTnull}
\alias{DTnull}
\title{Delete multiple columns from a data.table}
\usage{
DTnull(dt, cols)
}
\arguments{
\item{dt}{a data.table}

\item{cols}{a vector of column names}
}
\value{
a new data.table with specified columns deleted
}
\description{
Delete multiple columns from a data.table
}
