% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynDT.R
\name{dynDT}
\alias{dynDT}
\title{Create a dynamic data.}
\usage{
dynDT(x)
}
\arguments{
\item{x}{\code{[any]}}
}
\value{
\code{dynDT}: \code{[R6(dynDataTable)]}
}
\description{
table
}
\examples{
 dynDT(x = mtcars)
}
